Implement a Blacklist
This is a internal blacklist that only contains SteamID's of accounts that are either mine or are owned by friends of mine. Buying yourself into this blacklist is not possible.

If you don't want to use a Blacklist, pass the --noblacklist option with Titan when starting it from the command line.