Exit after the last bot successfully returned - resolves #76
This may not work as intended right now. The issue with Titan is that it is multi-threaded; all accounts are starting the botting at the same time and because of Steam the last account in the index is not guaranteed to exit as last one. I will address this in a future rework of the whole threading system done in Titan.